+++ app/src/main/java/woowacourse/shopping/view/cartcounter/OnClickCartItemCounter.kt
@@ -3,6 +3,6 @@ package woowacourse.shopping.view.cartcounter
 import woowacourse.shopping.domain.model.CartItemCounter
 
 interface OnClickCartItemCounter {
-    fun clickIncrease(productId:Int,itemPosition: Int, cartItemCounter: CartItemCounter)
-    fun clickDecrease(productId:Int,itemPosition: Int, cartItemCounter: CartItemCounter)
+    fun clickIncrease(productId:Long,itemPosition: Int, cartItemCounter: CartItemCounter)
+    fun clickDecrease(productId:Long,itemPosition: Int, cartItemCounter: CartItemCounter)
 }
